refactor: retire the skill-manifest lint with the corpus it validated (#17798)
The manifest and the corpus now live in neomjs/neo-agent-skills, and this
lint validated against a schema that repo deliberately narrowed — so it
was failing on fields the canonical schema no longer defines. A rule
enforced in N repositories is N places to drift; enforcement moved with
its subject rather than being duplicated behind it.

Ported to canonical first, each arm red on its own defect: manifest
schema validation and the growth-justification escape hatch. Without the
hatch the net-growth cap had no legitimate route past, which is how a
budget quietly stops being one.

parseSectionTriggers had one real importer — the Knowledge Base's
SkillSource — so it moved to its consumer instead of keeping a retired
lint alive for one function. That also removed an ai/services to
ai/scripts plane crossing, which is why one closure edge leaves the
receipt.

check-substrate-size loses its combined-budget arm: with the pr-review
files gone the array was empty, making the code path unreachable and
untestable rather than merely unused.

spec-retired: lintSkillManifest.spec.mjs retires with its subject — the rules it covered now live in neomjs/neo-agent-skills as scripts/lint-skill-corpus.mjs, proven by scripts/test-lint-skill-corpus.mjs (10 cases, each red on its own defect). Its one surviving neo behaviour, parseSectionTriggers, moved into ai/services/knowledge-base/source/SkillSource.mjs and is covered by SkillSource.spec.mjs. The combined-budget cases in CheckSubstrateSize.spec.mjs go the same way and for the same reason.
