+++ package.json
@@ -80,9 +80,7 @@
         "ai:lint-mcp-test-locations"           : "node ./ai/scripts/lint/lint-mcp-test-locations.mjs",
         "ai:lint-openapi-service-parity"       : "node ./ai/scripts/lint/lint-openapi-service-parity.mjs",
         "ai:lint-retry-bounds"                 : "node ./ai/scripts/lint/lint-retry-bounds.mjs",
-        "ai:lint-skill-manifest"               : "node ./ai/scripts/lint/lint-skill-manifest.mjs",
         "ai:lint-tree-json"                    : "node ./ai/scripts/lint/lint-tree-json.mjs",
-        "ai:skill-size-report"                 : "node ./ai/scripts/lint/lint-skill-manifest.mjs --report-sizes",
         "ai:kb-alerting"                       : "node ./ai/daemons/kb-alerting/daemon.mjs",
         "ai:kb-reconciliation"                 : "node ./ai/daemons/kb-reconciliation/daemon.mjs",
         "ai:kb-gc"                             : "node ./ai/daemons/kb-gc/daemon.mjs",
